requirements.txt
pyTelegramBotAPI==4.4.0
python-dotenv==0.20.0
loguru==0.6.0
PyMySQL==1.0.2
